+++ kamon-cassandra-client/src/test/scala/kamon/instrumentation/instrumentation/CassandraClientClientMetricsSpec.scala
@@ -18,7 +18,7 @@ package kamon.instrumentation.instrumentation
 import com.datastax.driver.core.Session
 import kamon.Kamon
 import kamon.instrumentation.cassandra.Cassandra.TargetNode
-import kamon.instrumentation.cassandra.{PoolMetrics, QueryMetrics}
+import kamon.instrumentation.cassandra.metrics.{PoolMetrics, QueryMetrics}
 import kamon.instrumentation.executor.ExecutorMetrics
 import kamon.tag.TagSet
 import kamon.testkit.{InstrumentInspection, MetricInspection}
